feat(app-auth): Standardize and Enhance Error handling Across Authentication flows

+++ src/services/auth/AuthService.ts
@@ -2,14 +2,13 @@ import { SignupSchema, SignupDto } from 'dtos/auth/SignupDto';
 import IAuthService from './interfaces/IAuthService';
 import IAuthRepository from 'repositories/auth/interfaces/IAuthRepository';
 import IHasher from 'types/utils/IHasher';
-import { ZodError } from 'zod';
 import { generateAccessToken, generateRefreshToken, verifyAccessToken } from 'utils/auth/tokenUtils';
 import RedisService from 'services/redis/RedisService';
 import IAuthServiceUser from './interfaces/IAuthUser';
 import { SigninDto, SigninSchema } from 'dtos/auth/SigninDto';
 import ITokenPayload from 'types/auth/ITokenPayload';
 import { ResetPasswordDto, ResetPasswordSchema } from 'dtos/auth/ResetPasswordDto';
-import { AuthenticationError, ServerError, ValidationError } from 'error/AppError';
+import { AppError, AuthenticationError, ServerError, ValidationError } from 'error/AppError';
 import { ErrorMessages } from 'constants/errorMessages';
 import { StatusCodes } from 'constants/statusCodes';
 
@@ -28,7 +27,7 @@ class AuthService implements IAuthService {
 
             const existingUser = await this._authRepository.findByPhoneNumber(signupData.phone_number);
             if (existingUser) {
-                throw new ValidationError(ErrorMessages.USER_ALREADY_EXISTS);
+                throw new ValidationError(ErrorMessages.USER_ALREADY_EXISTS, StatusCodes.BAD_REQUEST);
             }
             
             // Hash the password
@@ -43,29 +42,33 @@ class AuthService implements IAuthService {
             try {
                 accessToken = generateAccessToken(createUser);
                 refreshToken = generateRefreshToken(createUser);
-            } catch (tokenError) {
-                throw new ServerError(ErrorMessages.TOKEN_GENERATION_ERROR);
+            } catch (error) {
+                if (error) {
+                    throw new ServerError(ErrorMessages.TOKEN_GENERATION_ERROR, StatusCodes.INTERNAL_SERVER_ERROR);
+                }
             }
+
+            if (!accessToken || !refreshToken) throw new ServerError(ErrorMessages.TOKEN_GENERATION_ERROR, StatusCodes.INTERNAL_SERVER_ERROR);
 
             //Store the refresh token in Redis with a TTL 
             const REFRESH_TOKEN_TTL = 7 * 24 * 60 * 60;
             try {
                 await RedisService.storeRefreshToken(createUser.userId, refreshToken, REFRESH_TOKEN_TTL);
-            } catch (redisError) {
-                // Roll back user creation if Redis fails 
-                await RedisService.deleteRefreshToken(createUser.userId);
-                throw new ServerError(ErrorMessages.REFRESH_TOKEN_STORAGE_ERROR);
+            } catch (error) {
+                if (error) {
+                    // Roll back user creation if Redis fails 
+                    await RedisService.deleteRefreshToken(createUser.userId);
+                    throw new ServerError(ErrorMessages.REFRESH_TOKEN_STORAGE_ERROR, StatusCodes.INTERNAL_SERVER_ERROR);
+                }
             }
             
             // Send the response to the controler 
             return { ...createUser, accessToken };
         } catch (error) {
-            if (error instanceof ZodError) {
-                throw ErrorMessages.VALIDATION_ERROR;
-            } else if (error instanceof Error) {
-                throw error.message;
+            if (error instanceof AppError) {
+                throw error;
             } else {
-                throw new ServerError(ErrorMessages.INTERNAL_SERVER_ERROR);
+                throw error;
             }
         };
     } 
@@ -73,12 +76,18 @@ class AuthService implements IAuthService {
     async verifyToken(token: string): Promise<ITokenPayload> {
         try {
             const decodedData = verifyAccessToken(token);
+            if (!decodedData) throw new AuthenticationError(ErrorMessages.TOKEN_VERIFICATION_FAILED, StatusCodes.UNAUTHORIZED);
             const { phoneNumber } = decodedData;
             const getUserDetails = await this._authRepository.findByPhoneNumber(phoneNumber);
+            if (!getUserDetails) throw new ServerError(ErrorMessages.INTERNAL_SERVER_ERROR, StatusCodes.BAD_REQUEST);
             decodedData.status = getUserDetails?.status;
             return decodedData;
         } catch (error) {
-            throw new AuthenticationError(ErrorMessages.TOKEN_VERIFICATION_FAILED);
+            if (error instanceof AppError) {
+                throw error;
+            } else {
+                throw error;
+            }
         }
     }
 
@@ -106,68 +115,67 @@ class AuthService implements IAuthService {
             try {
                 accessToken = generateAccessToken(user);
                 refreshToken = generateRefreshToken(user);
-            } catch (tokenError) {
-                throw new ServerError(ErrorMessages.TOKEN_GENERATION_ERROR, StatusCodes.INTERNAL_SERVER_ERROR);
+            } catch (error) {
+                if (error) throw new ServerError(ErrorMessages.TOKEN_GENERATION_ERROR, StatusCodes.INTERNAL_SERVER_ERROR);
             }
+
+            if (!accessToken || !refreshToken) throw new ServerError(ErrorMessages.TOKEN_GENERATION_ERROR, StatusCodes.INTERNAL_SERVER_ERROR);
 
             //Store the refresh token in Redis with a TTL 
             const REFRESH_TOKEN_TTL = 7 * 24 * 60 * 60;
             try {
                 await RedisService.storeRefreshToken(user.userId, refreshToken, REFRESH_TOKEN_TTL);
-            } catch (redisError) {
-                throw new ServerError(ErrorMessages.REFRESH_TOKEN_STORAGE_ERROR, StatusCodes.INTERNAL_SERVER_ERROR);
+            } catch (error) {
+                if (error) throw new ServerError(ErrorMessages.REFRESH_TOKEN_STORAGE_ERROR, StatusCodes.INTERNAL_SERVER_ERROR);
             }
             
             // Send the response to the controller 
             return { ...user, accessToken };
         } catch (error) {
-            throw error;
+            if (error instanceof AppError) {
+                throw error;
+            } else {
+                throw error;
+            }
         };
     } 
 
     async signout(token: string): Promise<boolean> {
         try {
-            const verifiedUserDetails = this.verifyToken(token);
-            if (!verifiedUserDetails) {
-                throw new Error(`accessTokenVerification Failure`);
-            }
-            const userId = (await verifiedUserDetails).userId;
+            const verifiedUserDetails = await this.verifyToken(token);
+            const userId = verifiedUserDetails.userId;
 
             if (!userId) {
-                throw new Error(`User ID not found in verifiedUserDetails`);
+                throw new ServerError(ErrorMessages.STATUS_CHECK_FAILED, StatusCodes.INTERNAL_SERVER_ERROR);
             }
 
             const isRefreshTokenRemoved = await RedisService.deleteRefreshToken(userId);
 
             if (!isRefreshTokenRemoved!) {
-                throw new Error(`Failed to Remove refresh Token From Redis`);
+                throw new ServerError(ErrorMessages.REFRESH_TOKEN_REMOVAL_ERROR, StatusCodes.INTERNAL_SERVER_ERROR);
             }
             
             return true;
         } catch (error) {
-            // Extract and log the error message
-            let errorMessage = `An error occured while try to signout`;
-            if (error instanceof Error) {
-                errorMessage = error.message;
-            } 
-
-            throw new Error(errorMessage);
+            if (error instanceof AppError) {
+                throw error;
+            } else {
+                throw error;
+            }
         }
     }
 
     async verifyPhoneNumber(phoneNumber: string): Promise<boolean> {
         try {
-            if (!phoneNumber) throw new Error('Phone number is missing');
+            if (!phoneNumber) throw new ValidationError(ErrorMessages.PHONE_NUMBER_MISSING, StatusCodes.BAD_REQUEST);
             const userDetails = await this._authRepository.findByPhoneNumber(phoneNumber);
-            return userDetails?.status ?  true : false;
+            return !!userDetails?.status;
         } catch (error) {
-            // Extract and log the error message
-            let errorMessage = `Something went wrong!`;
-            if (error instanceof Error) {
-                errorMessage = error.message;
-            } 
-
-            throw new Error(errorMessage);
+            if (error instanceof AppError) {
+                throw error;
+            } else {
+                throw error;
+            }
         }
     }
 
@@ -178,7 +186,7 @@ class AuthService implements IAuthService {
 
             const user = await this._authRepository.findByPhoneNumber(data.phone_number);
             if (!user) {
-                throw new Error(`This number dosen't exists`);
+                throw new ValidationError(ErrorMessages.PHONE_NUMBER_NOT_FOUND, StatusCodes.UNAUTHORIZED);
             }
 
             const hashedPasswordInDatabase = user.hashedPassword;
@@ -188,7 +196,7 @@ class AuthService implements IAuthService {
             const isMatched = await this._hasher.verify(userProvidedPassword, hashedPasswordInDatabase!);
 
             if (isMatched) {
-                throw new Error("The entered password is the same as your current password. Please enter a new password.");
+                throw new ValidationError(ErrorMessages.PASSWORD_MATCH_ERROR, StatusCodes.BAD_REQUEST);
             }
 
              // Hash the password
@@ -198,15 +206,13 @@ class AuthService implements IAuthService {
             const userData = { ...data, password: hashedPassword };
 
             const isUpdated = await this._authRepository.resetPassword(userData);
-            return isUpdated ? true : false ;
+            return !!isUpdated;
         } catch (error) {
-            // Extract and log the error message
-            let errorMessage = `Something went wrong!`;
-            if (error instanceof Error) {
-                errorMessage = error.message;
-            } 
-
-            throw new Error(errorMessage);
+            if (error instanceof AppError) {
+                throw error;
+            } else {
+                throw error;
+            }
         }
     }
 }
